service fonksiyonları yazılıd

+++ src/main/java/com/workintech/model/repository/ActorRepository.java
@@ -3,6 +3,6 @@ package com.workintech.model.repository;
 import com.workintech.model.entity.Actor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ActorRepository extends JpaRepository<Actor,Integer> {
+public interface ActorRepository extends JpaRepository<Actor,Long> {
 
 }
